+++ packages/autocertifier-client/.eslintignore
@@ -1,4 +1,5 @@
 node_modules/**
 dist/**
 .idea/**
-src/proto/**
+generated/**
+
