+++ packages/autocertifier-client/src/exports.ts
@@ -15,6 +15,6 @@ export { InvalidSubdomainOrToken } from './errors'
 export { FailedToConnectToStreamrWebSocket } from './errors'
 export { ErrorCode } from './errors'
 export { Err } from './errors'
-export { HasSessionRequest, HasSessionResponse } from './proto/packages/autocertifier-client/protos/AutoCertifier'
-export { AutoCertifierRpcClient } from './proto/packages/autocertifier-client/protos/AutoCertifier.client'
+export { HasSessionRequest, HasSessionResponse } from '../generated/packages/autocertifier-client/protos/AutoCertifier'
+export { AutoCertifierRpcClient } from '../generated/packages/autocertifier-client/protos/AutoCertifier.client'
 export { createSelfSignedCertificate } from './createSelfSignedCertificate'
