+++ packages/autocertifier-client/tsconfig.browser.json
@@ -7,6 +7,7 @@
   },
   "include": [
     "src/**/*",
+    "generated/**/*",
     "test/**/*"
   ],
 }
