+++ packages/autocertifier-client/tsconfig.node.json
@@ -5,7 +5,8 @@
         "noImplicitOverride": false
     },
     "include": [
-        "src/**/*"
+        "src/**/*",
+        "generated/**/*"
     ],
     "references": [
         { "path": "../utils/tsconfig.node.json" },
